+++ lib/src/models/route_event.dart
@@ -39,7 +39,8 @@ class RouteEvent {
         eventType == MapBoxEvent.offline_region_progress ||
         eventType == MapBoxEvent.offline_region_complete ||
         eventType == MapBoxEvent.offline_region_error ||
-        eventType == MapBoxEvent.offline_region_removed) {
+        eventType == MapBoxEvent.offline_region_removed ||
+        eventType == MapBoxEvent.waypoint_arrival) {
       data = dataJson as Map<String, dynamic>?;
     } else if (eventType == MapBoxEvent.alternative_routes) {
       data = RouteAlternative.listFromJson(dataJson);